# dist/schemas/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LP REST API",
    "version": "1.0.0",
    "description": "REST API for Portfolio including Skills, Experience, and My Work"
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/experience": {
      "get": {
        "summary": "Get all experiences",
        "tags": [
          "Experience"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "expand",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Expand details by query params"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of experiences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "../doc/experiences.yaml#/components/schemas/Experience"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Add experience",
        "tags": [
          "Experience"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../doc/experiences.yaml#/components/schemas/ExperiencePost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "../doc/error.yaml#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "../doc/error.yaml#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "../doc/error.yaml#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/my-work": {
      "get": {
        "summary": "Get all my work",
        "tags": [
          "My Work"
        ],
        "responses": {
          "200": {
            "description": "A list of my work",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "../doc/myWork.yaml#/components/schemas/Work"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new work",
        "tags": [
          "My Work"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../doc/myWork.yaml#/components/schemas/WorkPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "../doc/error.yaml#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "../doc/error.yaml#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "../doc/error.yaml#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/skill": {
      "get": {
        "summary": "Get all skills",
        "tags": [
          "Skills"
        ],
        "responses": {
          "200": {
            "description": "A list of skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "../doc/skills.yaml#/components/schemas/Skill"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Add a new skill",
        "tags": [
          "Skills"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../doc/skills.yaml#/components/schemas/SkillPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "$ref": "../doc/error.yaml#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "../doc/error.yaml#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "../doc/error.yaml#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "../doc/error.yaml#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Contributor": {
        "type": "object",
        "required": [
          "name",
          "position",
          "communication"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the contributor"
          },
          "name": {
            "type": "string",
            "description": "Name of the contributor"
          },
          "position": {
            "type": "string",
            "description": "Position or role"
          },
          "photo": {
            "type": "string",
            "description": "Optional photo URL"
          },
          "communication": {
            "type": "string",
            "description": "Communication link or details"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "Experience": {
        "type": "object",
        "required": [
          "role",
          "orgName",
          "orgLogoUrl",
          "startDate",
          "endDate",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the experience"
          },
          "role": {
            "type": "string",
            "description": "Role or job title"
          },
          "orgName": {
            "type": "string",
            "description": "Name of the organization"
          },
          "orgLogoUrl": {
            "type": "string",
            "description": "URL to the organization's logo"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date of the experience"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "End date of the experience"
          },
          "description": {
            "type": "string",
            "description": "Details of the experience"
          }
        }
      },
      "ExperiencePost": {
        "type": "object",
        "required": [
          "role",
          "orgName",
          "orgLogoUrl",
          "startDate",
          "endDate",
          "description"
        ],
        "properties": {
          "role": {
            "type": "string",
            "description": "Role or job title"
          },
          "orgName": {
            "type": "string",
            "description": "Name of the organization"
          },
          "orgLogoUrl": {
            "type": "string",
            "description": "URL to the organization's logo"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Start date of the experience"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "End date of the experience"
          },
          "description": {
            "type": "string",
            "description": "Details of the experience"
          }
        }
      },
...
      "Work": {
        "type": "object",
        "required": [
          "title",
          "quickDoc",
          "category",
          "photoPrev",
          "creationDate",
          "doc",
          "technologies",
          "contributors",
          "src",
          "numOfImages",
          "images"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the work"
          },
          "title": {
            "type": "string",
            "description": "Title of the work"
          },
          "quickDoc": {
            "type": "string",
            "description": "Short documentation or description"
          },
          "category": {
            "type": "string",
            "enum": [
              "websites",
              "open-source",
              "ui-ux",
              "3d"
            ],
            "description": "Category of the work"
          },
          "photoPrev": {
            "type": "string",
            "description": "URL for the preview photo"
          },
          "creationDate": {
            "type": "string",
            "description": "Creation date of the work"
          },
          "doc": {
            "type": "string",
            "description": "Full documentation"
          },
          "technologies": {
            "type": "array",
            "description": "List of skills/technologies used",
            "items": {
              "$ref": "../doc/skills.yaml#/components/schemas/Skill"
            }
          },
          "contributors": {
            "type": "array",
            "description": "List of contributors",
            "items": {
              "$ref": "../doc/contributors.yaml#/components/schemas/Contributor"
            }
          },
          "src": {
            "type": "array",
            "description": "Source links",
            "items": {
              "type": "string"
            }
          },
          "numOfImages": {
            "type": "integer",
            "description": "Total number of images"
          },
          "images": {
            "type": "array",
            "description": "URLs for images",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WorkPost": {
        "type": "object",
        "required": [
          "title",
          "quickDoc",
          "category",
          "photoPrev",
          "creationDate",
          "doc",
          "technologies",
          "contributors",
          "src",
          "numOfImages",
          "images"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the work"
          },
          "quickDoc": {
            "type": "string",
            "description": "Short documentation or description"
          },
          "category": {
            "type": "string",
            "enum": [
              "websites",
              "open-source",
              "ui-ux",
              "3d"
            ],
            "description": "Category of the work"
          },
          "photoPrev": {
            "type": "string",
            "description": "URL for the preview photo"
          },
          "creationDate": {
            "type": "string",
            "description": "Creation date of the work"
          },
          "doc": {
            "type": "string",
            "description": "Full documentation"
          },
          "technologies": {
            "type": "array",
            "description": "List of skills/technologies used",
            "items": {
              "$ref": "../doc/skills.yaml#/components/schemas/Skill"
            }
          },
          "contributors": {
            "type": "array",
            "description": "List of contributors",
            "items": {
              "$ref": "../doc/contributors.yaml#/components/schemas/Contributor"
            }
          },
          "src": {
            "type": "array",
            "description": "Source links",
            "items": {
              "type": "string"
            }
          },
          "numOfImages": {
            "type": "integer",
            "description": "Total number of images"
          },
          "images": {
            "type": "array",
            "description": "URLs for images",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Skill": {
        "type": "object",
        "required": [
          "id",
          "title",
          "icon"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the skill"
          },
          "title": {
            "type": "string",
            "description": "Name of the skill"
          },
          "icon": {
            "type": "string",
            "description": "Icon URL or identifier for the skill"
          },
          "techUsage": {
            "type": "string",
            "description": "Optional technology usage details"
          }
        }
      },
      "SkillPost": {
        "type": "object",
        "required": [
          "title",
          "icon"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Name of the skill"
          },
          "icon": {
            "type": "string",
            "description": "Icon URL or identifier for the skill"
          },
          "techUsage": {
            "type": "string",
            "description": "Optional technology usage details"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}